requirements.txt
# Azure Functions runtime
azure-functions

# SQL Server connectivity
pyodbc

# Azure Identity and authentication
azure-identity
msal
msal-extensions

# Azure Key Vault
azure-keyvault-secrets

# Cryptography and security dependencies
cryptography
cffi
pycparser

# HTTP requests used by Azure SDKs
requests
requests-oauthlib
urllib3

# Optional: typing extensions
typing-extensions

# application insights

azure-functions
opencensus-ext-azure
opencensus-ext-logging
opencensus-ext-requests
pyodbc

